refactor(suite-native): simplify conditions in slice
