--- suite-native/device-authentication/src/index.ts
@@ -1,2 +1,0 @@
-export * from './deviceAuthenticationSlice';
-export * from './utils';
