+++ suite-native/device-authorization/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-native/device-authentication",
+    "name": "@suite-native/device-authorization",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
@@ -7,6 +7,7 @@
     "main": "src/index",
     "scripts": {
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "depcheck": "yarn g:depcheck",
         "type-check": "yarn g:tsc --build"
     },
     "dependencies": {
