+++ suite-native/device-authorization/src/deviceAuthorizationSlice.ts
@@ -1,0 +1,44 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+import { UI } from '@trezor/connect';
+
+import { isPinButtonRequestCode } from './utils';
+
+type DeviceAuthorizationState = {
+    hasDeviceRequestedPin: boolean;
+};
+
+type DeviceAuthorizationRootState = {
+    deviceAuthorization: DeviceAuthorizationState;
+};
+
+const deviceAuthorizationInitialState: DeviceAuthorizationState = {
+    hasDeviceRequestedPin: false,
+};
+
+export const deviceAuthorizationSlice = createSlice({
+    name: 'deviceAuthorization',
+    initialState: deviceAuthorizationInitialState,
+    reducers: {},
+    extraReducers: builder => {
+        builder
+            .addCase(UI.REQUEST_PIN, state => {
+                state.hasDeviceRequestedPin = true;
+            })
+            .addCase(UI.REQUEST_BUTTON, (state, action) => {
+                if (isPinButtonRequestCode(action)) {
+                    state.hasDeviceRequestedPin = true;
+                } else {
+                    state.hasDeviceRequestedPin = false;
+                }
+            })
+            .addCase(UI.CLOSE_UI_WINDOW, state => {
+                state.hasDeviceRequestedPin = false;
+            });
+    },
+});
+
+export const selectDeviceRequestedPin = (state: DeviceAuthorizationRootState) =>
+    state.deviceAuthorization.hasDeviceRequestedPin;
+
+export const deviceAuthorizationReducer = deviceAuthorizationSlice.reducer;
