+++ suite-native/device-authorization/src/index.ts
@@ -1,0 +1,2 @@
+export * from './deviceAuthorizationSlice';
+export * from './utils';
