+++ suite-native/device-authorization/tsconfig.json
