+++ suite-native/device/package.json
@@ -21,7 +21,7 @@
         "@suite-native/analytics": "workspace:*",
         "@suite-native/atoms": "workspace:*",
         "@suite-native/biometrics": "workspace:*",
-        "@suite-native/device-authentication": "workspace:*",
+        "@suite-native/device-authorization": "workspace:*",
         "@suite-native/device-mutex": "workspace:*",
         "@suite-native/feature-flags": "workspace:*",
         "@suite-native/intl": "workspace:*",
