+++ suite-native/device/src/hooks/useHandleDeviceConnection.ts
@@ -20,7 +20,7 @@ import {
     selectIsDeviceUsingPassphrase,
     authorizeDeviceThunk,
 } from '@suite-common/wallet-core';
-import { selectDeviceRequestedPin } from '@suite-native/device-authentication';
+import { selectDeviceRequestedPin } from '@suite-native/device-authorization';
 import { selectIsOnboardingFinished } from '@suite-native/settings';
 import { requestPrioritizedDeviceAccess } from '@suite-native/device-mutex';
 import { useIsBiometricsOverlayVisible } from '@suite-native/biometrics';
