+++ suite-native/device/src/middlewares/buttonRequestMiddleware.ts
@@ -1,7 +1,6 @@
 import { createMiddlewareWithExtraDeps } from '@suite-common/redux-utils';
 import { UI } from '@trezor/connect';
 import { deviceActions, selectDevice } from '@suite-common/wallet-core';
-import { isPinAction } from '@suite-native/device-authentication';
 
 export const prepareButtonRequestMiddleware = createMiddlewareWithExtraDeps(
     (action, { dispatch, getState, next }) => {
@@ -18,8 +17,7 @@ export const prepareButtonRequestMiddleware = createMiddlewareWithExtraDeps(
             );
         }
 
-        // Pin is not handled by buttonRequest but redux state within @suite-native/device-authentication
-        if (action.type === UI.REQUEST_BUTTON && !isPinAction(action)) {
+        if (action.type === UI.REQUEST_BUTTON) {
             const { device: _, ...request } = action.payload;
 
             dispatch(
