+++ suite-native/device/tsconfig.json
@@ -12,7 +12,7 @@
         { "path": "../analytics" },
         { "path": "../atoms" },
         { "path": "../biometrics" },
-        { "path": "../device-authentication" },
+        { "path": "../device-authorization" },
         { "path": "../device-mutex" },
         { "path": "../feature-flags" },
         { "path": "../intl" },
