+++ suite-native/module-connect-device/package.json
@@ -20,7 +20,7 @@
         "@suite-native/alerts": "workspace:*",
         "@suite-native/atoms": "workspace:*",
         "@suite-native/device": "workspace:*",
-        "@suite-native/device-authentication": "workspace:*",
+        "@suite-native/device-authorization": "workspace:*",
         "@suite-native/device-mutex": "workspace:*",
         "@suite-native/forms": "workspace:*",
         "@suite-native/intl": "workspace:*",
