+++ suite-native/state/package.json
@@ -25,7 +25,7 @@
         "@suite-native/blockchain": "workspace:*",
         "@suite-native/config": "workspace:*",
         "@suite-native/device": "workspace:*",
-        "@suite-native/device-authentication": "workspace:*",
+        "@suite-native/device-authorization": "workspace:*",
         "@suite-native/discovery": "workspace:*",
         "@suite-native/feature-flags": "workspace:*",
         "@suite-native/graph": "workspace:*",
