+++ suite-native/state/tsconfig.json
@@ -28,7 +28,7 @@
         { "path": "../blockchain" },
         { "path": "../config" },
         { "path": "../device" },
-        { "path": "../device-authentication" },
+        { "path": "../device-authorization" },
         { "path": "../discovery" },
         { "path": "../feature-flags" },
         { "path": "../graph" },
