+++ tsconfig.json
@@ -54,7 +54,7 @@
         { "path": "suite-native/config" },
         { "path": "suite-native/device" },
         {
-            "path": "suite-native/device-authentication"
+            "path": "suite-native/device-authorization"
         },
         { "path": "suite-native/device-manager" },
         { "path": "suite-native/device-mutex" },
